fix tree shaking
